Move OverrideWidget to config_utils.py
