+++ .github/wordlist.txt
@@ -1,9 +1,7 @@
-autofill
 Bing
 Bool
-bool
 Combobox
-df
+CheckBoxWidget
 DataFrame
 DAC
 GeoDataFrame
@@ -12,27 +10,32 @@ Gurobi
 HBox
 IntSlider
 Jupyter
-libs
 Mcf
 Namespace
 Pyomo
 Pytest
 Quickstart
+SubCheckBoxWidget
 Redistributions
 arg
 argparse
+autofill
+bool
 checkboxes
 codebase
 conda
 config
 csv
 customizable
+df
 env
 geopoint
 geospatial
 gpd
 ipynb
+ipywidgets
 json
+libs
 ndarray
 np
 num
