+++ pytest.ini
@@ -1,5 +1,6 @@
 [pytest]
 testpaths = primo/
 markers =
-    secrets: tests that rely on repository secrets which may not be accesible to PRs
     demo: test that relies on running the demo notebook
+    secrets: tests that rely on repository secrets which may not be accesible to PRs
+    widgets: tests for widgets that rely on playwright and solara for execution
